Create AddContact component.
- Add create contact button
- Renamed contact search bar to Header
- Seperate AddContact component from Header compoent
